backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
tensorflow==2.15.0
Pillow==10.3.0
numpy==1.26.4
opencv-python==4.9.0.80
reportlab==4.2.0
google-generativeai==0.7.0
scikit-learn==1.5.0
matplotlib==3.8.4
pandas==2.2.2
mlflow==2.14.1
dvc[gdrive]==3.50.1
python-dotenv==1.0.1
pytest==8.2.2
gunicorn==22.0.0